backend/requirements.txt
Django==3.2.7
django-extensions==3.1.3
django-polymorphic==3.1.0
django-crum==0.7.9
django-filter==2.4.0
django-guardian==2.4.0
django-mptt==0.13.3
django-nose==1.4.7
django-redis==4.12.1
django-simple-captcha==0.5.14
argon2-cffi==21.1.0

eulxml==1.1.3
isodate==0.6.0
lxml==4.6.3

nose==1.3.7
Pillow==8.3.2
psycopg2-binary==2.9.1

requests==2.26.0
tablib==3.0.0
gunicorn==20.1.0
uvicorn==0.15.0

djangorestframework==3.12.4
djangorestframework-gis==0.17
djangorestframework-jsonapi==4.2.1
djangorestframework-jsonapi[django-filter]==4.2.1
drf-extensions==0.7.1
uritemplate==4.1.1

redis==3.5.3
celery==5.0.5
django-celery-beat==2.2.0
django-celery-results==2.2.0

websockets==9.1
channels==3.0.4
channels-redis==3.3.0